pyproject.toml
[project]
name = "ckb-qa-tool-v0-1-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.14",
    "cryptography>=45.0.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pyinstaller>=6.14.2",
    "toml>=0.10.2",
    "websockets>=15.0.1",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
